split model files
