+++ iOS/Museek/Museek/Band.swift
@@ -31,26 +31,4 @@ extension Band: JSONTransformable {
     
 }
 
-class Event: Model {
-    
-    let id = Property<String>(key: "id", required: true)
-    let name = Property<String>(key: "name", required: true)
-    let date = Property<NSDate>(key: "date", required: true)
-    let venue = PropertyDictionary<String>(key: "venue", required: true)
-    let URLs = PropertyArray<NSURL>(key: "urls")
-    
-    let band = Property<Band>(key: "band", required: true)
 
-}
-
-extension Event: MKAnnotation {
-    
-    var coordinate: CLLocationCoordinate2D {
-        return CLLocationCoordinate2D(latitude: 0, longitude: 0)
-    }
-    
-    var title: String? {
-        return name.value
-    }
-    
-}
